test(milkie): daemon e2e — 池 spawn+create_agent+run_turn+shutdown 子进程退出;配置示例
